requirements/requirements.txt
# mycroft
HolmesV[all]>=2021.5.6a14
mock_msm
mycroft-messagebus-client~=0.9,!=0.9.2

# neon stuff
neon_speech @ git+https://github.com/NeonGeckoCom/neon_speech
neon_audio @ git+https://github.com/NeonGeckoCom/neon_audio
neon_enclosure @ git+https://github.com/NeonGeckoCom/neon_enclosure

# utils
rapidfuzz
kthread
# TODO needs 0.0.12 once released
# see https://github.com/OpenVoiceOS/ovos_utils/pull/74
ovos_utils>=0.0.11
ovos-skills-manager>=0.0.2

# plugins
ovos-plugin-manager>=0.0.1a2
libretranslate_neon_plugin

# text parser modules
RAKEkeywords>=0.2.0